Update Github actions flow
This takes the lessons learned while implementing it. It now has a
single workflow file for all tests in all events. Caching should also be
correct now.

A major benefit is that the unit and acceptance test matrices are now
generated dynamically at runtime, based on what's in metadata.json. That
means less churn in changes to supported versions.

It also does a very basic validation prior to generating the matrices.
This ensures that no expensive test matrix is started on invalid syntax.