Use gha-puppet's reusable workflows
This changes the CI to use gha-puppet. The goal of gha-puppet is to
avoid templating huge workflows. Rather it should be short.

It changes PUPPET_VERSION to PUPPET_GEM_VERSION to be compatible with
the action. This is done to align with the larger ecosystem.
PUPPET_VERSION is also used by beaker-puppet_install_helper but with a
different syntax so it always conflicted.

For now Rubocop is turned off since none of our modules are ready for
it.

The install_epel variable is removed because since
56e6607759f818eb31d6f7fea94050bed9d63977 it no longer exists.

The locale_workaround is introduced for modules that use PostgreSQL on
EL7.